python: configables of any type
A Configable subclass no longer needs to be instantiated with a
dictionary. For example, a complex Configable object could be the
result of a detailed string.

Also, SUBTYPE class member can now be a staticmethod that accepts
the initializing value and returns True if the class should be
instantiated instead of the parent type.